+++ apps/server/__tests__/api/admin/login.test.ts
@@ -1,9 +1,9 @@
-import controller from '$/api/login/admin/controller';
+import controller from '$/api/admin/login/controller';
 import { createAdminAuthorizationPayload } from '$/service/auth';
 import { getAdmin, validateUser } from '$/repositories/admin';
 import { createHash } from '$/repositories/utils';
 import { getFastify, getPromiseLikeItem } from '$/__tests__/utils';
-import { login } from '$/domains/login/admin';
+import { login } from '$/domains/admin/login';
 
 describe('login', () => {
   it('should login', async() => {
@@ -39,6 +39,7 @@ describe('login', () => {
     })(getFastify());
 
     const res = await injectedController.post({
+      user: { id: 1, roles: [] },
       body: {
         email: 'test@example.com',
         pass: 'test1234',
@@ -80,6 +81,7 @@ describe('login', () => {
     })(getFastify());
 
     const res = await injectedController.post({
+      user: { id: 1, roles: [] },
       body: {
         email: 'test@example.com',
         pass: 'wrong pass',
