+++ apps/server/__tests__/service/auth.test.ts
@@ -2,6 +2,12 @@
 import { verifyAdmin, getRolesValue, verifyGuest } from '$/service/auth';
 
 describe('verifyAdmin', () => {
+  it('should return true if login api', async() => {
+    const jwtVerify = jest.fn();
+    expect(await verifyAdmin({ url: '/api/login', method: 'POST', jwtVerify } as any)).toBe(true);
+    expect(jwtVerify).not.toBeCalled();
+  });
+
   it('should return true if verified', async() => {
     const jwtVerify = jest.fn(() => ({
       id: 123,
