+++ apps/server/api/admin/login/controller.ts
@@ -1,5 +1,5 @@
 import { defineController } from './$relay';
-import { login } from '$/domains/login/admin';
+import { login } from '$/domains/admin/login';
 
 export default defineController(({ login }), ({ login }, fastify) => ({
   post: ({ body }) => login(body, fastify),
