+++ apps/server/api/front/login/controller.ts
@@ -1,5 +1,5 @@
 import { defineController } from './$relay';
-import { login } from '$/domains/login/front';
+import { login } from '$/domains/front/login';
 
 export default defineController(({ login }), ({ login }, fastify) => ({
   post: ({ body }) => login(body, fastify),
