--- apps/server/api/login/controller.ts
@@ -1,3 +1,0 @@
-import { defineController } from './$relay';
-
-export default defineController(() => ({}));
