--- apps/server/api/login/index.ts
@@ -1,1 +1,0 @@
-export type Methods = {}
