+++ apps/server/domains/admin/validators.ts
@@ -3,3 +3,4 @@ export * from './guests/validators';
 export * from './rooms/validators';
 export * from './reservations/validators';
 export * from './admins/validators';
+export * from './login/validators';
