+++ apps/server/domains/front/validators.ts
@@ -1,2 +1,3 @@
 export * from './account/validators';
 export * from './reservation/validators';
+export * from './login/validators';
