--- apps/server/domains/login/validators.ts
@@ -1,2 +1,0 @@
-export * from './admin/validators';
-export * from './front/validators';
