+++ apps/server/domains/validators.ts
@@ -1,4 +1,3 @@
 export * from './admin/validators';
 export * from './front/validators';
-export * from './login/validators';
 export * from './lock/validators';
